Use gevent.event.Event() to signal greenlet completion
This way we only enter the idle loop in one place, meaning
we can inline it in run().
It would be nice if we could avoid the timeout in wait(),
but it seems urwid relies on the mainloop to initiate a
rendering run, so we have to iterate every second.